requirements.txt
Flask
Flask_SQLAlchemy
psycopg2-binary
PyJWT
Flask-Bcrypt
python-dotenv

